LLM_Playground/llm_playground_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0


